some methods were fixed + new tests were added.
